requirements.txt
requests
python-dotenv
sqlalchemy
psycopg2-binary
pandas
streamlit
